rough refactoring
