+++ geoparser/app/templates/annotate.html
